% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_from_gps.R
\name{block_from_gps}
\alias{block_from_gps}
\title{Extract CDFW block id from GPS coordinates}
\usage{
block_from_gps(lat_dd, long_dd)
}
\arguments{
\item{lat_dd}{Latitude in decimal degrees}

\item{long_dd}{Longitude in decimal degrees}
}
\value{
Block id
}
\description{
This function extracts CDFW block ids from a list of GPS coordinates.
}
\examples{
# Extract block ids
# The second to last coordinate is purposefully incomplete
# The last coordinate is purposefully outside the block area
lat_dd <- c(33.338592, 35.551473, 36.716016, NA, 36.716016)
long_dd <- c(-118.310696, -121.108906, -121.820242, -121.820242, -130)
block_from_gps(lat_dd=lat_dd, long_dd=long_dd)
}
