% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calcom}
\alias{calcom}
\title{CALCOM annual groundfish landings by port complex and gear type}
\format{
A data frame with the following attributes:
\describe{
  \item{market_group}{Species market category (e.g., rockfish, flatfish, elasmobranch, etc.)}
  \item{species_code}{Species code}
  \item{common_name}{Species common name}
  \item{scientific_name}{Species scientific name}
  \item{port_complex_code}{Port complex code}
  \item{port_complex}{Port complex}
  \item{gear_code}{Gear type code}
  \item{gear}{Gear type (e.g., hook and line, net, traps, trawl, etc.)}
  \item{year}{Year}
  \item{source}{Source}
  \item{live_yn}{Live (yes or no)}
  \item{landings_lb}{Landings volume in pounds (lb)}
}
}
\usage{
calcom
}
\description{
Annual landings of California groundfish by port complex, gear type, and species from 1978-2019 based on the CALCOM database.
}
\keyword{datasets}
