% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_cpfv}
\alias{cdfw_cpfv}
\title{CDFW annual landings from Commercial Passenger Fishing Vessels}
\format{
A data frame with the following attributes:
\describe{
  \item{...}{Descriptions coming soon}
}
}
\usage{
cdfw_cpfv
}
\description{
Annual landings from Commercial Passenger Fishing Vessels (CPFVs) in California from 1936-2019.
}
\keyword{datasets}
