% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_cpfv_effort}
\alias{cdfw_cpfv_effort}
\title{CDFW annual participation in and effort by Commercial Passenger Fishing Vessels}
\format{
A data frame with the following attributes:
\describe{
  \item{source_landings}{Source and table for landings data}
  \item{source_cpfvs}{Source and table for registered vessel data}
  \item{source_anglers}{Source and table for licensed angler data}
  \item{source_time}{Source and table for angler effort data}
  \item{region}{Regional scale (statewide or northern/southern)}
  \item{port_complex_group}{Port complex, 1987-2019 typology}
  \item{port_complex}{Port sub-complex, as in the original data}
  \item{year}{Year}
  \item{landings_n}{Number of landed fish}
  \item{cpfvs_n}{Number of registered CPFVs}
  \item{anglers_n}{Number of CPFV anglers}
  \item{days_n}{Number of days fished by CPFV anglers}
  \item{hours_n}{Number of hours fished by CPFV anglers}
}
}
\usage{
cdfw_cpfv_effort
}
\description{
Annual participation in and effort by from Commercial Passenger Fishing Vessels (CPFVs) in California from 1936-2019.
}
\keyword{datasets}
