% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_kelp}
\alias{cdfw_kelp}
\title{CDFW annual harvest of kelp by bed type}
\format{
A data frame with the following attributes:
\describe{
  \item{source}{Source (Fish Bulletin or CDFW report)}
  \item{table}{Table number (e.g., Table 4, Table 5, etc.)}
  \item{year}{Year}
  \item{open_bed_t}{Harvest (short tons) from open beds}
  \item{leased_bed_t}{Harvest (short tons) from leased beds}
  \item{total_bed_t}{Total harvest (short tons) - sum of open and leased harvest}
}
}
\usage{
cdfw_kelp
}
\description{
Annual harvest of kelp (Macrocystis spp.) by bed type (open vs. leased) from 1916-1976. Harvest is reported in "short" tons (2000 lbs).
}
\keyword{datasets}
