% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_kelp_tots}
\alias{cdfw_kelp_tots}
\title{CDFW annual harvest of kelp}
\format{
A data frame with the following attributes:
\describe{
  \item{source}{Source (Fish Bulletin, NOAA FOSS, or CDFW website)}
  \item{year}{Year}
  \item{type}{Reported or Estimated (using DataThief)}
  \item{harvest_lb}{Total harvest in pounds (lbs)}
  \item{harvest_t}{Total harvest in short tons (t) - a short ton equals 2000 lbs)}
  \item{harvest_kg}{Total harvest in kilograms (kg)}
  \item{harvest_mt}{Total harvest in metric tons (mt)}
}
}
\usage{
cdfw_kelp_tots
}
\description{
Annual harvest of kelp (Macrocystis spp.) from 1916-2019. The data for 2003-2019 was extracted from the figure on the following web-page using DataThief: https://wildlife.ca.gov/Conservation/Marine/Kelp/Commercial-Harvest
}
\keyword{datasets}
