% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_n_comm_fishers}
\alias{cdfw_n_comm_fishers}
\title{CDFW annual number of licensed commercial fishers}
\format{
A data frame with the following attributes:
\describe{
  \item{...}{Descriptions coming soon}
}
}
\usage{
cdfw_n_comm_fishers
}
\description{
Annual number of licensed commercial fishers in California from 1916-2020 with information by area of residence from 1935-1976.
}
\keyword{datasets}
