% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_n_comm_vessels_length}
\alias{cdfw_n_comm_vessels_length}
\title{CDFW annual number of registered commercial fishing vessels by length class}
\format{
A data frame with the following attributes:
\describe{
  \item{...}{Descriptions coming soon}
}
}
\usage{
cdfw_n_comm_vessels_length
}
\description{
Annual number of registered commercial fishing vessels in California by length class from 1934-1956.
}
\keyword{datasets}
