% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_n_comm_vessels_port}
\alias{cdfw_n_comm_vessels_port}
\title{CDFW annual number of registered commercial fishing vessels by port complex}
\format{
A data frame with the following attributes:
\describe{
  \item{source}{Source (Fish Bulletin or CDFW report)}
  \item{table}{Table number (e.g., Table 4, Table 5, etc.)}
  \item{season}{License year, April 1st to March 31st}
  \item{year}{Year}
  \item{port_complex}{Port complex (e.g., San Francisco, Monterey, etc.)}
  \item{nvessels}{Number of registered commercial fishing vessels}
}
}
\usage{
cdfw_n_comm_vessels_port
}
\description{
Annual number of registered commercial fishing vessels in California by port complex from 1934-1976.
}
\keyword{datasets}
