% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_ports}
\alias{cdfw_ports}
\title{CDFW annual commercial fisheries landings by port and species}
\format{
A data frame with the following attributes:
\describe{
  \item{source_type}{Fish Bulletin or CDFW website}
  \item{source}{Source of data}
  \item{table}{Table containing data in source}
  \item{port_complex}{Port complex, based on the most recept typology}
  \item{port_complex_orig}{Port complex, as in original data}
  \item{port}{Port name, harmonized across years}
  \item{port_orig}{Port name, as in original data}
  \item{type}{Landings or shipments}
  \item{year}{Year}
  \item{comm_name}{Common name, harmonized across years and datasets}
  \item{comm_name_orig}{Common name, as in original data}
  \item{sci_name}{Scientific name}
  \item{presentation}{Presentation (i.e., roe, roe on kelp, claws)}
  \item{value_used}{Ex-vessel value (US dollars)}
  \item{landings_lb}{Landings in pounds}
  \item{landings_kg}{Landings in kilograms}
}
}
\usage{
cdfw_ports
}
\description{
Annual California commercial fisheries landings and shipments (pounds/values) by port and species from 1941-2019.
}
\keyword{datasets}
