% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_waters}
\alias{cdfw_waters}
\title{CDFW annual commercial fisheries landings by waters and species}
\format{
A data frame with the following attributes:
\describe{
  \item{source}{Source of data}
  \item{table}{Table containing data in source}
  \item{year}{Year}
  \item{waters}{Waters (e.g., California, north-of-state, south-of-state, etc.)}
  \item{comm_name_orig}{Common name, as in original data}
  \item{comm_name}{Common name, harmonized across years and datasets}
  \item{sci_name}{Scientific name}
  \item{presentation}{Presentation (i.e., roe, roe on kelp, claws)}
  \item{landings_lb}{Landings in pounds}
  \item{landings_kg}{Landings in kilograms}
}
}
\usage{
cdfw_waters
}
\description{
Annual California commercial fisheries landings and shipments (pounds) by waters and species from 1936-2019.
}
\keyword{datasets}
