% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_names.R
\name{harmonize_names}
\alias{harmonize_names}
\title{Harmonize scientific and common names for West Coast fish species}
\usage{
harmonize_names(x, from, to)
}
\arguments{
\item{x}{A character vector of species common names}

\item{from}{Type of data being harmonized}

\item{to}{Type of data to return}
}
\value{
A character vector of harmonized names
}
\description{
This function harmonizes the scientific and common names for West Coast fish species.
}
\examples{
# Suggest spelling corrections for common names
harmonize_names(x=c("Pacific hake", "Northern halibut"), from="comm", to="comm")
harmonize_names(x=c("Dungeness crab", "Northern anchovy"), from="comm", to="sci")
harmonize_names(x=c("Metacarcinus magister", "Engraulis mordax"), from="sci", to="comm")
}
