% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loran_to_gps.R
\name{loran_to_gps}
\alias{loran_to_gps}
\title{Convert Loran-C coordinates to GPS coordinates}
\usage{
loran_to_gps(chain, loran_x, loran_w)
}
\arguments{
\item{chain}{Chain}

\item{loran_x}{Loran-C x value}

\item{loran_w}{Loran-C w value}
}
\value{
Data frame of GPS values
}
\description{
This function converts Loran-C coordinates to GPS coordinates by using a ghost browser to run this NOAA conversion tool: https://apps-nefsc.fisheries.noaa.gov/loran/public/
}
\details{
The chain must be one of the following 4-digit numbers: 5930 (Canadian E. Coast), 5990 (Canadian W. Coast), 7270 (Newfoundland E. Coast), 7960 (Gulf of Alaska), 7980 (Southeast U.S.), 8290 (N. Central U.S.), 8970 (Great Lakes), 9610 (S. Central U.S.), 9940 (Western U.S.), 9960 (Northeast U.S.), 9990 (N. Pacific).
}
\examples{
# Convert Loran coordinates
# The second to last value is purposefully invalid
loran_x <- c(27138, 27142, 27144, 27144)
loran_w <- c(15338, 14327, 153366, 15336)
loran_to_gps(chain=9940, loran_x=loran_x, loran_w=loran_w)
}
