% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa}
\alias{noaa}
\title{NOAA annual commercial fisheries landings by state and species}
\format{

}
\usage{
noaa
}
\description{
An 'sf' object (similar to a shapefile) mapping the West Coast's commercial fishing blocks.
}
\keyword{datasets}
