% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_all1}
\alias{pacfin_all1}
\title{PACFIN annual commercial fisheries landings by species}
\format{
A data frame with the following attributes:
\describe{
  \item{year}{Year}
  \item{state}{State}
  \item{......}{More coming soon}
}
}
\usage{
pacfin_all1
}
\description{
Annual commercial fisheries landings (pounds/value) by species from 1980-2024 from the PACFIN database.
}
\keyword{datasets}
