% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_all2}
\alias{pacfin_all2}
\title{PACFIN annual commercial fisheries landings by species and gear}
\format{
A data frame with the following attributes:
\describe{
  \item{state}{State}
  \item{mgmt_group}{Management group}
  \item{complex}{Complex}
  \item{comm_name}{Common name}
  \item{spp_code}{PACFIN species code}
  \item{gear}{Gear}
  \item{year}{Year}
  \item{catch_mt}{Catch (mt)}
  \item{revenue_usd}{Revenues (USD)}
  \item{price_usd_lb}{Price (USD/lb)}
  \item{confidential}{Confidential}
}
}
\usage{
pacfin_all2
}
\description{
Annual commercial fisheries landings (pounds/value) by species and gear from 1980-2022 from the PACFIN database.
}
\keyword{datasets}
