% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_all5}
\alias{pacfin_all5}
\title{PACFIN annual commercial fisheries landings by port complex and species}
\format{
A data frame with the following attributes:
\describe{
  \item{year}{Year}
  \item{state}{State}
  \item{......}{More coming soon}
}
}
\usage{
pacfin_all5
}
\description{
Annual commercial fisheries landings (pounds/value) by port complex (CA/OR/WA) and species from 1980-2020 from the PACFIN database.
}
\keyword{datasets}
