% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_crab2}
\alias{pacfin_crab2}
\title{PACFIN monthly commercial Dungeness crab landings by port complex}
\format{
A data frame with the following attributes:
\describe{
  \item{season}{Season}
  \item{year}{Year}
  \item{month}{Month}
  \item{date}{Date}
  \item{state}{State}
  \item{port_code}{Port complex code}
  \item{port_complex}{Port complex name}
  \item{confidential}{Confidentiality flag}
  \item{landings_mt}{Landings (mt)}
  \item{price_usd_kg}{Price ($/lb)}
  \item{revenues_usd}{Revenues (USD)}
}
}
\usage{
pacfin_crab2
}
\description{
Monthly commercial Dungeness crab landings (pounds/value) by port complex (CA/OR/WA) from 1980-2020 from the PACFIN database.
}
\keyword{datasets}
