% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_ports}
\alias{pacfin_ports}
\title{PACFIN ports meta-data}
\format{
A data frame with the following attributes:
\describe{
  \item{port_code}{Port code}
  \item{state1}{State name}
  \item{state2}{State abbreviation}
  \item{port_name}{Port name}
  \item{port_yn}{Port (yes/no)}
  \item{long_dd}{Longitude (°W)}
  \item{lat_dd}{Latitude (°N)}
}
}
\usage{
pacfin_ports
}
\description{
Ports in the PACFIN database.
}
\keyword{datasets}
