% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_species}
\alias{pacfin_species}
\title{PACFIN species meta-data}
\format{
A data frame with the following attributes:
\describe{
  \item{spp_code}{Species code}
  \item{comm_name}{Common name}
  \item{sci_name}{Scientific name}
  \item{mgmt_group_code}{Management group code}
  \item{complex_code}{Species complex code}
}
}
\usage{
pacfin_species
}
\description{
Species in the PACFIN database.
}
\keyword{datasets}
