% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot data from the CALFISH database}
\usage{
plot_data(dataset)
}
\arguments{
\item{dataset}{Name of one of the CALFISH database datasets (lowercase with underscores)}
}
\value{
A plot of one of the CALFISH database datasets
}
\description{
Prepares a quick plot of the datasets published in the CALFISH database.
}
\examples{
# Plot CDFW datasets quickly
plot_data(dataset="cdfw_waters")
plot_data(dataset="cdfw_ports")
plot_data(dataset="cdfw_n_comm_fishers")
plot_data(dataset="cdfw_n_comm_vessels_length")
plot_data(dataset="cdfw_n_comm_vessels_port")
plot_data(dataset="cdfw_n_comm_vessels")
plot_data(dataset="cdfw_cpfv")
plot_data(dataset="cdfw_cpfv_effort")
plot_data(dataset="cdfw_kelp")
plot_data(dataset="cdfw_kelp_tots")
}
