% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ports}
\alias{ports}
\title{California's commercial fishing ports}
\format{
A data frame with 143 rows (ports) and 11 variables (port attributes):
\describe{
  \item{state}{State}
  \item{port_complex_orig}{Port complex(s) used in the original data}
  \item{port_complex1}{Port complex, 1987-2019 typologogy}
  \item{port_complex2}{Port complex, 1936-1986 typologogy}
  \item{port_complex3}{Port complex, 1931-1935 typologogy}
  \item{port_complex4}{Port complex, 1926-1930 typologogy}
  \item{county}{County}
  \item{port}{Port name}
  \item{type}{Individual port or group of ports}
  \item{lat_dd}{Latitude (°N)}
  \item{long_dd}{Longitude (°W)}
}
}
\usage{
ports
}
\description{
Identity and location of California's commercial fishing ports.
}
\keyword{datasets}
