% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recfin_cte2}
\alias{recfin_cte2}
\title{RECFIN annual recreational landings and discards by state and more}
\format{
A data frame with the following attributes:
\describe{
  \item{state}{State}
  \item{water_area}{Water area}
  \item{mode}{Mode}
  \item{trip_type}{Trip type}
  \item{comm_name}{Common name}
  \item{year}{Year}
  \item{status}{Retained or released (dead)}
  \item{catch_n}{Catch, in number of fish}
  \item{catch_mt}{Catch, in metric tons}
}
}
\usage{
recfin_cte2
}
\description{
Annual recreational fisheries landings and discards by state (CA/OR/WA), water area, mode, trip type, and species from 2001-2020 from the RECFIN database.
}
\keyword{datasets}
