% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recfin_cte3}
\alias{recfin_cte3}
\title{RECFIN annual recreational mortality by state and mode}
\format{
A data frame with the following attributes:
\describe{
  \item{state}{State}
  \item{mode}{Mode}
  \item{comm_name}{Common name}
  \item{year}{Year}
  \item{catch_n}{Catch, in number of fish}
  \item{catch_mt}{Catch, in metric tons}
}
}
\usage{
recfin_cte3
}
\description{
Annual recreational fisheries mortality by state (CA/OR/WA), mode, and species from 2001-2020 from the RECFIN database.
}
\keyword{datasets}
