% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recfin_cte5}
\alias{recfin_cte5}
\title{RECFIN annual recreational salmon landings and discards from OR and WA}
\format{
A data frame with the following attributes:
\describe{
  \item{year}{Year}
  \item{month}{Month}
  \item{week}{Week}
  \item{state}{State}
  \item{district}{District}
  \item{mode}{Mode}
  \item{water_area}{Water area}
  \item{subregion}{Subregion}
  \item{trip_type}{Trip type}
  \item{comm_name}{Common name}
  \item{sci_name}{Scientific name}
  \item{level}{Level (species or group)}
  \item{retained_n}{Number of retained fish}
  \item{released_n}{Number of released fish}
}
}
\usage{
recfin_cte5
}
\description{
Annual recreational salmon landings and discards from OR and WA from 2001-2020 from the RECFIN database.
}
\keyword{datasets}
