% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recfin_cte7}
\alias{recfin_cte7}
\title{RECFIN annual recreational salmon landings and discards from CA}
\format{
A data frame with the following attributes:
\describe{
  \item{year}{Year}
  \item{month}{Month}
  \item{agency_code}{Agency code}
  \item{agency}{Agency}
  \item{state}{State}
  \item{area}{area}
  \item{water_area_code}{Water area code}
  \item{water_area}{Water area}
  \item{mode}{Mode}
  \item{spp_code}{Species code}
  \item{comm_name}{Common name}
  \item{sci_name}{Scientific name}
  \item{level}{Level (species or group)}
  \item{retained_n}{Number of retained fish}
}
}
\usage{
recfin_cte7
}
\description{
Annual recreational salmon landings and discards from CA from 1976-2021 from the RECFIN database.
}
\keyword{datasets}
