% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recfin_species}
\alias{recfin_species}
\title{RECFIN species key}
\format{
A data frame with the following attributes:
\describe{
  \item{comm_name_orig}{Common name in original data}
  \item{comm_name}{Common name, formatted}
  \item{sci_name}{Scientific name, as judged by Chris Free}
}
}
\usage{
recfin_species
}
\description{
An unofficial (i.e., made by Chris Free) key to species in the RECFIN data.
}
\keyword{datasets}
