% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_names.R
\name{reverse_names}
\alias{reverse_names}
\title{Reverse common names}
\usage{
reverse_names(names)
}
\arguments{
\item{x}{Common name}
}
\value{
Reversed common name
}
\description{
This function reverses fish common name. For example, it changes "Halibut, Pacific" to "Pacific halibut" and vice versa.
}
\examples{
# Reverse common name
reverse_names(names=c("California spiny lobster", "Pacific halibut", "Hake"))
reverse_names(names=c("Lobster, California spiny", "Halibut, Pacific", "Hake", "Crab, rock"))
}
