% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swfsc}
\alias{swfsc}
\title{SWFSC monthly landings from California waters by port complex and species}
\format{
A data frame with the following attributes:
\describe{
  \item{dataset}{Dataset}
  \item{port_complex}{Port complex}
  \item{year}{Year}
  \item{comm_name_orig}{Common name, as in the original data}
  \item{sci_name}{Scientific name}
  \item{level}{Taxonomic resolution (species- or group-specific)}
  \item{presentation}{Presentation (a.k.a., condition) of the landings (i.e., roe, roe on kelp, claws, not specified)}
  \item{landings_lb}{Landings in pounds (lbs)}
  \item{landings_kg}{Landings in kilograms (kg)}
}
}
\usage{
swfsc
}
\description{
Monthly landings from California waters by port complex and species from 1928-2002 based on the Mason et al. (2004) digitization effort.
}
\keyword{datasets}
