% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa}
\alias{taxa}
\title{West Coast fisheries species information}
\format{
A data frame with the following attributes::
\describe{
  \item{comm_name}{Common name}
  \item{scientific_name}{Scientific name}
  \item{type}{Type (e.g., fish, crustacean, mollusk, other invertebrate)}
  \item{taxa_group}{Taxonomic group}
  \item{phylum}{Phylum}
  \item{class}{Class}
  \item{order}{Order}
  \item{family}{Family}
  \item{genus}{Genus}
}
}
\usage{
taxa
}
\description{
Taxonomic information for West Coast fisheries species.
}
\keyword{datasets}
