% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blocks}
\alias{blocks}
\title{West Coast commercial fishing blocks}
\format{
An 'sf' object data frame with 836 rows (blocks) and 7 variables (block attributes):
\describe{
  \item{block_state}{Block state (California, Oregon, or Washington)}
  \item{block_type}{Block type (inshore, midshore, offshore)}
  \item{block_id}{Block id}
  \item{block_sqkm}{Block area (sq. km)}
  \item{block_lat_dd}{Latitude (°N) of centroid}
  \item{block_long_dd}{Longitude (°W) of centroid}
}
}
\usage{
blocks
}
\description{
An 'sf' object (similar to a shapefile) mapping the West Coast's commercial fishing blocks.
}
\keyword{datasets}
