% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_cpfv}
\alias{cdfw_cpfv}
\title{CDFW annual landings from Commercial Passenger Fishing Vessels}
\format{
A data frame with the following attributes:
\describe{
  \item{waters}{Waters (Fish Bulletin or CDFW report)}
  \item{comm_name}{Species common name, harmonized across datasets}
  \item{comm_name_orig}{Species common name, as in the original data}
  \item{sci_name}{Species scientific name}
  \item{level}{Level (species or group)}
  \item{year}{Year}
  \item{landings_n}{Landings in number of fish}
}
}
\usage{
cdfw_cpfv
}
\description{
Annual landings from Commercial Passenger Fishing Vessels (CPFVs) in California from 2000-2019.
}
\keyword{datasets}
