% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_kelp}
\alias{cdfw_kelp}
\title{CDFW annual harvest of kelp by bed type}
\format{

}
\usage{
cdfw_kelp
}
\description{
Annual harvest of kelp (Macrocystis spp.) by bed type (open vs. leased) from 1916-1976. Harvest is reported in "short" tons (2000 lbs).
}
\keyword{datasets}
