% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_n_comm_fishers}
\alias{cdfw_n_comm_fishers}
\title{CDFW annual number of licensed commercial fishers}
\format{
A data frame with the following attributes:
\describe{
  \item{source}{Source (Fish Bulletin or CDFW report)}
  \item{table}{Table number (e.g., Table 4, Table 5, etc.)}
  \item{season}{License year, April 1st to March 31st}
  \item{year}{Year}
  \item{region_type}{Scale of regional information (statewide or area-of-residence)}
  \item{region_group}{Region group (other categories grouped together)}
  \item{region}{Region (statewide or name or port complex)}
  \item{nfishers}{Total number of licensed commercial fishers}
  \item{nfishers_nr}{Number of non-resident licensed commercial fishers}
  \item{nfishers_r}{Number of resident licensed commercial fishers}
}
}
\usage{
cdfw_n_comm_fishers
}
\description{
Annual number of licensed commercial fishers in California from 1916-2020 with information by area of residence from 1935-1976.
}
\keyword{datasets}
