% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_n_comm_vessels_port}
\alias{cdfw_n_comm_vessels_port}
\title{CDFW annual number of registered commercial fishing vessels by port complex}
\format{
A data frame with the following attributes:
\describe{
  \item{...}{Descriptions coming soon}
}
}
\usage{
cdfw_n_comm_vessels_port
}
\description{
Annual number of registered commercial fishing vessels in California by port complex from 1934-1976.
}
\keyword{datasets}
