% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_ports}
\alias{cdfw_ports}
\title{CDFW annual commercial fisheries landings by port and species}
\format{
A data frame with the following attributes:
\describe{
  \item{year}{Year}
  \item{......}{More coming soon}
}
}
\usage{
cdfw_ports
}
\description{
Annual California commercial fisheries landings and shipments (pounds/values) by port and species from 1941-2019.
}
\keyword{datasets}
