% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfw_waters}
\alias{cdfw_waters}
\title{CDFW annual commercial fisheries landings by waters and species}
\format{
A data frame with the following attributes:
\describe{
  \item{year}{Year}
  \item{state}{State}
  \item{......}{More coming soon}
}
}
\usage{
cdfw_waters
}
\description{
Annual California commercial fisheries landings and shipments (pounds) by waters and species from 1936-2019.
}
\keyword{datasets}
