% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa}
\alias{noaa}
\title{NOAA annual commercial fisheries landings by state and species}
\format{

}
\usage{
noaa
}
\description{
Annual commercial fisheries landings (pounds/value) by U.S. state and species from 1950-2019 from NOAA's FOSS database.
}
\keyword{datasets}
